requirements.txt
dash==2.0.0
gunicorn==20.0.1
pandas==2.0.1
pillow==9.5.0
pyarrow==12.0.0